Add function to default to sample files if settings file not found
